# execbuilder: add "average lookup ratio" parallelization heuristic for lookup joins