--- a/pkg/ccl/logictestccl/testdata/logic_test/regional_by_row_cascade
+++ b/pkg/ccl/logictestccl/testdata/logic_test/regional_by_row_cascade
@@ -58,15 +58,36 @@ INSERT INTO child (c, p, crdb_region) VALUES (1000, 100, 'us-east-1'), (2000, 20
 statement ok
 ANALYZE great_grandparent;
 
+# Only the scan in the main query is parallelized when we don't have stats on
+# the descendant tables.
+query I
+SELECT count(*) FROM [EXPLAIN (VERBOSE) DELETE FROM great_grandparent WHERE i = 1] WHERE info LIKE '%parallel%';
+----
+1
+
 statement ok
 ANALYZE grandparent;
+
+# Now we also should parallelize lookup join into the grandparent table.
+query I
+SELECT count(*) FROM [EXPLAIN (VERBOSE) DELETE FROM great_grandparent WHERE i = 1] WHERE info LIKE '%parallel%';
+----
+2
 
 statement ok
 ANALYZE parent;
 
+# Now we also should parallelize lookup join into the parent table.
+query I
+SELECT count(*) FROM [EXPLAIN (VERBOSE) DELETE FROM great_grandparent WHERE i = 1] WHERE info LIKE '%parallel%';
+----
+3
+
 statement ok
 ANALYZE child;
 
+# Finally, all three lookup joins as well as the scan in the main query should
+# be parallelized.
 query T
 EXPLAIN (VERBOSE) DELETE FROM great_grandparent WHERE i = 1;
 ----
@@ -126,6 +147,7 @@ vectorized: true
         │               │ estimated row count: 3
         │               │ table: grandparent@grandparent_gg_idx
         │               │ lookup condition: (crdb_region IN ('ap-southeast-2', 'ca-central-1', 'us-east-1')) AND (gg = gg)
+        │               │ parallel
         │               │
         │               └── • distinct
         │                   │ columns: (gg)
@@ -163,6 +185,7 @@ vectorized: true
                 │               │ estimated row count: 3
                 │               │ table: parent@parent_g_idx
                 │               │ lookup condition: (crdb_region IN ('ap-southeast-2', 'ca-central-1', 'us-east-1')) AND (g = g)
+                │               │ parallel
                 │               │
                 │               └── • distinct
                 │                   │ columns: (g)
@@ -193,6 +216,7 @@ vectorized: true
                                 │ estimated row count: 3
                                 │ table: child@child_p_idx
                                 │ lookup condition: (crdb_region IN ('ap-southeast-2', 'ca-central-1', 'us-east-1')) AND (p = p)
+                                │ parallel
                                 │
                                 └── • distinct
                                     │ columns: (p)
@@ -206,3 +230,106 @@ vectorized: true
                                               columns: (p, g, crdb_region)
                                               estimated row count: 100
                                               label: buffer 1000000
+
+statement ok
+SET parallelize_multi_key_lookup_joins_avg_lookup_ratio = 0;
+
+# Only the scan in the main query is parallelized when the "average lookup
+# ratio" heuristic is disabled.
+query I
+SELECT count(*) FROM [EXPLAIN (VERBOSE) DELETE FROM great_grandparent WHERE i = 1] WHERE info LIKE '%parallel%';
+----
+1
+
+statement ok
+RESET parallelize_multi_key_lookup_joins_avg_lookup_ratio;
+
+# All three lookup joins as well as the scan in the main query should be
+# parallelized.
+query I
+SELECT count(*) FROM [EXPLAIN (VERBOSE) DELETE FROM great_grandparent WHERE i = 1] WHERE info LIKE '%parallel%';
+----
+4
+
+# Inject the table stats for grandparent table to simulate the case when each
+# region stores 100k rows each. The lookup into the table should still be
+# parallelized (if it's not, then we're using the wrong ColumnIDs when
+# retrieving column stats).
+statement ok
+ALTER TABLE grandparent INJECT STATISTICS '[
+    {
+        "avg_size": 4,
+        "columns": [
+            "crdb_region"
+        ],
+        "created_at": "2025-01-01 00:00:00.000000",
+        "distinct_count": 3,
+        "histo_col_type": "",
+        "name": "__auto__",
+        "null_count": 0,
+        "row_count": 300000
+    },
+    {
+        "avg_size": 2,
+        "columns": [
+            "gg"
+        ],
+        "created_at": "2025-01-01 00:00:00.000000",
+        "distinct_count": 300000,
+        "histo_buckets": [
+            {"distinct_range": 0, "num_eq": 1, "num_range": 0, "upper_bound": "1"},
+            {"distinct_range": 299999, "num_eq": 1, "num_range": 299999, "upper_bound": "300000"}
+        ],
+        "histo_col_type": "INT8",
+        "histo_version": 3,
+        "name": "__auto__",
+        "null_count": 0,
+        "row_count": 300000
+    }
+]'
+
+query I
+SELECT count(*) FROM [EXPLAIN (VERBOSE) DELETE FROM great_grandparent WHERE i = 1] WHERE info LIKE '%parallel%';
+----
+4
+
+# Now simulate a scenario where many rows have NULLs in the lookup column 'gg'.
+# The lookup into the table should still be parallelized (if it's not, then
+# we're incorrectly considering NULLs in the heuristic).
+statement ok
+ALTER TABLE grandparent INJECT STATISTICS '[
+    {
+        "avg_size": 4,
+        "columns": [
+            "crdb_region"
+        ],
+        "created_at": "2025-01-01 00:00:00.000000",
+        "distinct_count": 3,
+        "histo_col_type": "",
+        "name": "__auto__",
+        "null_count": 0,
+        "row_count": 1000000
+    },
+    {
+        "avg_size": 2,
+        "columns": [
+            "gg"
+        ],
+        "created_at": "2025-01-01 00:00:00.000000",
+        "distinct_count": 300000,
+        "histo_buckets": [
+            {"distinct_range": 0, "num_eq": 1, "num_range": 0, "upper_bound": "1"},
+            {"distinct_range": 299999, "num_eq": 1, "num_range": 299999, "upper_bound": "300000"}
+        ],
+        "histo_col_type": "INT8",
+        "histo_version": 3,
+        "name": "__auto__",
+        "null_count": 700000,
+        "row_count": 1000000
+    }
+]'
+
+query I
+SELECT count(*) FROM [EXPLAIN (VERBOSE) DELETE FROM great_grandparent WHERE i = 1] WHERE info LIKE '%parallel%';
+----
+4