--- a/pkg/ccl/logictestccl/testdata/logic_test/regional_by_row_query_behavior
+++ b/pkg/ccl/logictestccl/testdata/logic_test/regional_by_row_query_behavior
@@ -3582,7 +3582,7 @@ ALTER TABLE xyz INJECT STATISTICS '[
 
 # Regression test for #105942.
 query T retry
-EXPLAIN SELECT
+EXPLAIN (VERBOSE) SELECT
   xyz.str,
   abc.id,
   abc.id1,
@@ -3603,20 +3603,28 @@ WHERE
 distribution: local
 vectorized: true
 ·
-• lookup join
-│ estimated row count: 1
-│ table: xyz@xyz_id2_str_abc_id_idx
-│ equality: (crdb_region, id2) = (crdb_region, id2)
-│ pred: ((abc_id = id) AND (id2 = '68088706-02c6-47d1-b993-a421cd761f2b')) AND (crdb_region = 'ap-southeast-2')
+• project
+│ columns: (str, id, id1, id2, created_at, updated_at)
 │
-└── • index join
-    │ estimated row count: 0
-    │ table: abc@abc_pkey
+└── • lookup join (inner)
+    │ columns: (id, id1, created_at, updated_at, id2, crdb_region, str, abc_id, id2, crdb_region)
+    │ estimated row count: 1
+    │ table: xyz@xyz_id2_str_abc_id_idx
+    │ equality: (crdb_region, id2) = (crdb_region, id2)
+    │ pred: ((abc_id = id) AND (id2 = '68088706-02c6-47d1-b993-a421cd761f2b')) AND (crdb_region = 'ap-southeast-2')
     │
-    └── • scan
-          estimated row count: 0 (0.37% of the table; stats collected <hidden> ago)
-          table: abc@abc_id1_id2_idx
-          spans: [/'ap-southeast-2'/'6da4f356-e526-4b78-b9f9-bbb1a7fc12d6'/'68088706-02c6-47d1-b993-a421cd761f2b' - /'ap-southeast-2'/'6da4f356-e526-4b78-b9f9-bbb1a7fc12d6'/'68088706-02c6-47d1-b993-a421cd761f2b']
+    └── • index join
+        │ columns: (id, id1, created_at, updated_at, id2, crdb_region)
+        │ estimated row count: 0
+        │ table: abc@abc_pkey
+        │ key columns: crdb_region, id
+        │ parallel
+        │
+        └── • scan
+              columns: (id, id1, id2, crdb_region)
+              estimated row count: 0 (0.37% of the table; stats collected <hidden> ago)
+              table: abc@abc_id1_id2_idx
+              spans: /"@"/"m\xa4\xf3V\xe5&Kx\xb9\xf9\xbb\xb1\xa7\xfc\x12\xd6"/"h\b\x87\x06\x02\xc6Gѹ\x93\xa4!\xcdv\x1f+"-/"@"/"m\xa4\xf3V\xe5&Kx\xb9\xf9\xbb\xb1\xa7\xfc\x12\xd6"/"h\b\x87\x06\x02\xc6Gѹ\x93\xa4!\xcdv\x1f+"/PrefixEnd
 
 # The following should use a string of 4 lookup/index joins with a cost under 200.
 query T retry