--- a/pkg/sql/exec_util.go
+++ b/pkg/sql/exec_util.go
@@ -3871,6 +3871,22 @@ func (m *sessionDataMutator) SetParallelizeMultiKeyLookupJoinsEnabled(val bool) 
 	m.data.ParallelizeMultiKeyLookupJoinsEnabled = val
 }
 
+func (m *sessionDataMutator) SetParallelizeMultiKeyLookupJoinsAvgLookupRatio(val float64) {
+	m.data.ParallelizeMultiKeyLookupJoinsAvgLookupRatio = val
+}
+
+func (m *sessionDataMutator) SetParallelizeMultiKeyLookupJoinsMaxLookupRatio(val float64) {
+	m.data.ParallelizeMultiKeyLookupJoinsMaxLookupRatio = val
+}
+
+func (m *sessionDataMutator) SetParallelizeMultiKeyLookupJoinsAvgLookupRowSize(val int64) {
+	m.data.ParallelizeMultiKeyLookupJoinsAvgLookupRowSize = val
+}
+
+func (m *sessionDataMutator) SetParallelizeMultiKeyLookupJoinsOnlyOnMRMutations(val bool) {
+	m.data.ParallelizeMultiKeyLookupJoinsOnlyOnMRMutations = val
+}
+
 // TODO(harding): Remove this when costing scans based on average column size
 // is fully supported.
 func (m *sessionDataMutator) SetCostScansWithDefaultColSize(val bool) {