--- a/pkg/sql/logictest/testdata/logic_test/pg_catalog
+++ b/pkg/sql/logictest/testdata/logic_test/pg_catalog
@@ -3083,7 +3083,11 @@ optimizer_use_provided_ordering_fix                              on             
 optimizer_use_trigram_similarity_optimization                    on                  NULL      NULL        NULL        string
 optimizer_use_virtual_computed_column_stats                      on                  NULL      NULL        NULL        string
 override_multi_region_zone_config                                off                 NULL      NULL        NULL        string
+parallelize_multi_key_lookup_joins_avg_lookup_ratio              10                  NULL      NULL        NULL        string
+parallelize_multi_key_lookup_joins_avg_lookup_row_size           100 KiB             NULL      NULL        NULL        string
 parallelize_multi_key_lookup_joins_enabled                       off                 NULL      NULL        NULL        string
+parallelize_multi_key_lookup_joins_max_lookup_ratio              10000               NULL      NULL        NULL        string
+parallelize_multi_key_lookup_joins_only_on_mr_mutations          on                  NULL      NULL        NULL        string
 password_encryption                                              scram-sha-256       NULL      NULL        NULL        string
 pg_trgm.similarity_threshold                                     0.3                 NULL      NULL        NULL        string
 plan_cache_mode                                                  auto                NULL      NULL        NULL        string
@@ -3320,7 +3324,11 @@ optimizer_use_provided_ordering_fix                              on             
 optimizer_use_trigram_similarity_optimization                    on                  NULL  user     NULL      on                  on
 optimizer_use_virtual_computed_column_stats                      on                  NULL  user     NULL      on                  on
 override_multi_region_zone_config                                off                 NULL  user     NULL      off                 off
+parallelize_multi_key_lookup_joins_avg_lookup_ratio              10                  NULL  user     NULL      10                  10
+parallelize_multi_key_lookup_joins_avg_lookup_row_size           100 KiB             B     user     NULL      100 KiB             100 KiB
 parallelize_multi_key_lookup_joins_enabled                       off                 NULL  user     NULL      off                 off
+parallelize_multi_key_lookup_joins_max_lookup_ratio              10000               NULL  user     NULL      10000               10000
+parallelize_multi_key_lookup_joins_only_on_mr_mutations          on                  NULL  user     NULL      on                  on
 password_encryption                                              scram-sha-256       NULL  user     NULL      scram-sha-256       scram-sha-256
 pg_trgm.similarity_threshold                                     0.3                 NULL  user     NULL      0.3                 0.3
 plan_cache_mode                                                  auto                NULL  user     NULL      auto                auto
@@ -3548,7 +3556,11 @@ optimizer_use_provided_ordering_fix                              NULL    NULL   
 optimizer_use_trigram_similarity_optimization                    NULL    NULL     NULL     NULL        NULL
 optimizer_use_virtual_computed_column_stats                      NULL    NULL     NULL     NULL        NULL
 override_multi_region_zone_config                                NULL    NULL     NULL     NULL        NULL
+parallelize_multi_key_lookup_joins_avg_lookup_ratio              NULL    NULL     NULL     NULL        NULL
+parallelize_multi_key_lookup_joins_avg_lookup_row_size           NULL    NULL     NULL     NULL        NULL
 parallelize_multi_key_lookup_joins_enabled                       NULL    NULL     NULL     NULL        NULL
+parallelize_multi_key_lookup_joins_max_lookup_ratio              NULL    NULL     NULL     NULL        NULL
+parallelize_multi_key_lookup_joins_only_on_mr_mutations          NULL    NULL     NULL     NULL        NULL
 password_encryption                                              NULL    NULL     NULL     NULL        NULL
 pg_trgm.similarity_threshold                                     NULL    NULL     NULL     NULL        NULL
 plan_cache_mode                                                  NULL    NULL     NULL     NULL        NULL