--- a/pkg/sql/logictest/testdata/logic_test/show_source
+++ b/pkg/sql/logictest/testdata/logic_test/show_source
@@ -187,7 +187,11 @@ optimizer_use_provided_ordering_fix                              on
 optimizer_use_trigram_similarity_optimization                    on
 optimizer_use_virtual_computed_column_stats                      on
 override_multi_region_zone_config                                off
+parallelize_multi_key_lookup_joins_avg_lookup_ratio              10
+parallelize_multi_key_lookup_joins_avg_lookup_row_size           100 KiB
 parallelize_multi_key_lookup_joins_enabled                       off
+parallelize_multi_key_lookup_joins_max_lookup_ratio              10000
+parallelize_multi_key_lookup_joins_only_on_mr_mutations          on
 password_encryption                                              scram-sha-256
 pg_trgm.similarity_threshold                                     0.3
 plan_cache_mode                                                  auto