… pkg/ccl/logictestccl/tests/multiregion-9node-3region-3azs/BUILD.bazel
… pkg/ccl/logictestccl/tests/multiregion-9node-3region-3azs/generated_test.go
… pkg/sql/opt/exec/execbuilder/testdata/lookup_join
… pkg/sql/opt/lookupjoin/constraint_builder.go
… pkg/sql/opt/props/histogram_test.go
… pkg/sql/opt/xform/join_funcs.go
… pkg/sql/opt/xform/testdata/rules/join
… pkg/sql/recursive_cte.go
… pkg/sql/sessiondatapb/local_only_session_data.proto
… pkg/sql/vars.go