@yuzefovich 2025-07-01
This is RFAL.

There are a few TODOs in the code that I'm curious to hear thoughts on. I also plan to do some more empirical testing to fine-tune the defaults but that doesn't need to block the review.