@michae2 2025-07-09
@rytaft talked me down a bit, and pointed out this is an innovation release, which is a great time to gather data about the efficacy of the heuristic. So I take it back about shipping it disabled. I'd be ok turning it on for some set of statements.