@yuzefovich 2025-07-09
Pushed an update which makes the new heuristic only apply to mutations of multi-region tables by default (it can be applied to all stmts by setting new session variable `parallelize_multi_key_lookup_joins_only_on_mr_mutations` to `false`). Also reverted some of the test changes that were related to read-only stmts or mutations of non-multi-region tables.