@yuzefovich 2025-07-10
Newly added logging exposed incorrect context capture during recursive CTE iterations ("span use after finish"), so I added another commit to fix that.