@yuzefovich 2025-07-17
It's pretty random, but for posterity the main commit in this PR results in a microbenchmark regression in pgwire:
```
args: benchdiff "--run=^BenchmarkWriteBinaryColumnarDecimal" "--old=9f84aa54^" "--new=9f84aa54" "./pkg/sql/pgwire" "-b"

name                           old time/op    new time/op    delta
WriteBinaryColumnarDecimal-24     299µs ± 1%     387µs ± 1%  +29.39%  (p=0.000 n=9+10)

name                           old alloc/op   new alloc/op   delta
WriteBinaryColumnarDecimal-24    49.1kB ± 0%    49.1kB ± 0%     ~     (p=0.246 n=10+9)

name                           old allocs/op  new allocs/op  delta
WriteBinaryColumnarDecimal-24     2.05k ± 0%     2.05k ± 0%     ~     (all equal)
```